проделнна неплохая работа добавил напоминания, контейнезировал сам бот, поменял промт, и стиль пару сообщений

+++ bot/handlers/load_every_day_report_handler.py
@@ -7,7 +7,7 @@ from aiogram.fsm.context import FSMContext
 from aiogram.filters.state import State, StatesGroup,StateFilter
 from api_openrouter.api_openrouter import get_ai
 from promts.every_day_report import promt_every_day_report
-from database.queries import load_today_to_db
+from database.queries import load_today_to_db,has_report_today
 from psycopg import AsyncConnection
 
 main_router = Router()
@@ -22,12 +22,15 @@ async def start_process(message: Message):
                          reply_markup= get_callback_inline_keyboard('Ежедневный отчет', 'Мои отчеты'))
 
 @main_router.callback_query(F.data == 'Ежедневный отчет')
-async def every_day_report(callback_query: CallbackQuery,state: FSMContext):
-    await callback_query.message.answer(
-        'Отлично бро!\n Давай с тобой обсудим новый день. Для начала дай ему оценку от 1 до 10. Стоп слово - ВСЕ')
-    await callback_query.message.delete()
-    await state.clear()
-    await state.set_state(UserStates.mood)
+async def every_day_report(callback_query: CallbackQuery,state: FSMContext,conn: AsyncConnection):
+    if await has_report_today(conn, callback_query.from_user.id):
+        await callback_query.message.answer('❗️ За сегодня уже есть отчет, так что давай дождемся завтрашнего дня!')
+    else:
+        await callback_query.message.answer(
+            'Отлично бро!\n Давай с тобой обсудим новый день. Для начала дай ему оценку от 1 до 10. Стоп слово - ВСЕ')
+        await callback_query.message.delete()
+        await state.clear()
+        await state.set_state(UserStates.mood)
 
 
 @main_router.message(lambda x: x.text.isdigit() and 0<int(x.text)<=10,StateFilter(UserStates.mood) )
@@ -47,6 +50,7 @@ async def get_mood_fail(message: Message, state: FSMContext):
     else:
         await state.clear()
         await message.answer('Заполение отчета сброшено!')
+        await start_process(message)
 
 @main_router.message(lambda x: len(x.text.strip())>50, StateFilter(UserStates.day_desc))
 async def get_day_desc(message: Message, state: FSMContext):
@@ -82,6 +86,7 @@ async def get_day_desc_fail(message: Message, state: FSMContext):
 @main_router.message(StateFilter(UserStates.conclusion))
 async def get_conclusion(message: Message, state: FSMContext, conn:AsyncConnection):
     data = await state.get_data()
-    await load_today_to_db(conn, data['mood'], data['day_desc'], data['mentor_desc'],message.text)
-    await message.answer('Спасибо за твой отчет! Надеюсь ты вынес из этого свои выводы. Этот отчет сохарнился, ты можешь посмотреть его в лююое время')
+    await load_today_to_db(conn,message.from_user.id, data['mood'], data['day_desc'], data['mentor_desc'],message.text)
+    await message.answer('Спасибо за твой отчет! Надеюсь ты вынес из этого свои выводы. Этот отчет сохарнился, ты можешь посмотреть его в любое время')
     await state.clear()
+    await start_process(message)
